厨房原型文件

+++ SunacCADApp/Controllers/SysRoleController.cs
@@ -13,8 +13,14 @@ namespace SunacCADApp.Controllers
     public class SysRoleController : Controller
     {
 
+        private string UserId = string.Empty;
+        private string UserName = string.Empty;
+        private int IUserId = 0;
         public SysRoleController() 
         {
+            IUserId = InitUtility.Instance.InitSessionHelper.Get("UserID").ConvertToInt32(0);
+            UserId = InitUtility.Instance.InitSessionHelper.Get("UserID");
+            UserName = InitUtility.Instance.InitSessionHelper.Get("UserName");
             ViewBag.SelectModel = 12;
         }
 
@@ -27,6 +33,10 @@ namespace SunacCADApp.Controllers
         [ValidateInput(false)]
         public ActionResult Index()
         {
+            if (IUserId < 1)
+            {
+                return Redirect("/home");
+            }
             string _where = "1=1";  //查询
             string _orderby = string.Empty;  //排序
             string _url ="1&";
@@ -65,6 +75,10 @@ namespace SunacCADApp.Controllers
         [ValidateInput(false)]
         public ActionResult Add()
         {
+            if (IUserId < 1)
+            {
+                return Redirect("/home");
+            }
             string _wh = "1=1";
             string _ordby = " Id asc,";
             IList<Sys_Model> models = Sys_ModelDB.GetPageInfoByParameter(_wh, _ordby, 0, 30);
@@ -80,21 +94,28 @@ namespace SunacCADApp.Controllers
         [ValidateInput(false)]
         public ActionResult Addhandle()
         {
+            if (IUserId < 1)
+            {
+                return Redirect("/home");
+            }
             string role_name = Request.Form["txt_role_name"].ConventToString(string.Empty);
              Sys_Role hasRole =  SysRoleDB.GetSingleEntityByparam(string.Format(" And  Role_Name='{0}'",role_name));
              if (hasRole != null) 
              {
                  return Json(new { code = -100, message = role_name+"角色已存在" }, JsonRequestBehavior.AllowGet);
              }
-
             Sys_Role sys_role = new Sys_Role();
             sys_role.Role_Name = Request.Form["txt_role_name"].ConventToString(string.Empty);
             sys_role.Role_Remark = Request.Form["area_role_remark"].ConventToString(string.Empty);
             sys_role.CreateOn = DateTime.Now;
             sys_role.Reorder = Request.Form["txt_reorder"].ConvertToInt32(0);
-            sys_role.Enabled = Request.Form["select_enabled"].ConvertToInt32(0);
-            sys_role.CreateUserId = 0;
-            sys_role.CreateBy = "admin";
+            sys_role.Enabled = 1;
+            sys_role.CreateUserId = UserId.ConvertToInt32(0);
+            sys_role.CreateBy = UserName;
+            sys_role.CreateOn = DateTime.Now;
+            sys_role.ModifiedUserId = UserId.ConvertToInt32(0);
+            sys_role.ModifiedBy = UserName;
+            sys_role.ModifiedOn = DateTime.Now;
             int roleid = SysRoleDB.AddHandle(sys_role);
             string powers = Request.Form["hid_power"].ConventToString(string.Empty);
             if (!string.IsNullOrEmpty(powers)) 
@@ -105,9 +126,13 @@ namespace SunacCADApp.Controllers
                     Sys_Role_Model_Relation RoleModel = new Sys_Role_Model_Relation();
                     RoleModel.Model_Id = power.ConvertToInt32(-1);
                     RoleModel.Role_Id = roleid;
-                    RoleModel.CreateOn = DateTime.Now;
                     RoleModel.Enabled = 1;
                     RoleModel.Reorder = 0;
+                    RoleModel.CreateUserId = UserId.ConvertToInt32(0);
+                    RoleModel.CreateBy = UserName;
+                    RoleModel.CreateOn = DateTime.Now;
+                    RoleModel.ModifiedUserId = UserId.ConvertToInt32(0);
+                    RoleModel.ModifiedBy = UserName;
                     RoleModel.ModifiedOn = DateTime.Now;
                     SysRoleModelRelationDB.AddHandle(RoleModel);
                 }
@@ -130,6 +155,10 @@ namespace SunacCADApp.Controllers
         /// <author>alon<84789887@qq.com></author>  
         public ActionResult Edit(int Id)
         {
+            if (IUserId < 1)
+            {
+                return Redirect("/home");
+            }
             string _wh = "1=1";
             string _ordby = " Id asc,";
             IList<Sys_Model> models = Sys_ModelDB.GetPageInfoByParameter(_wh, _ordby, 0, 30);
@@ -151,6 +180,10 @@ namespace SunacCADApp.Controllers
         [ValidateInput(false)]
         public ActionResult Edithandle()
         {
+            if (IUserId < 1)
+            {
+                return Redirect("/home");
+            }
             Sys_Role sys_role = new Sys_Role();
             int Id = Request.Form["hid_id"].ConvertToInt32(0);
             sys_role.Id = Id;
@@ -158,9 +191,10 @@ namespace SunacCADApp.Controllers
             sys_role.Role_Remark = Request.Form["area_role_remark"].ConventToString(string.Empty);
             sys_role.CreateOn = DateTime.Now;
             sys_role.Reorder = Request.Form["txt_reorder"].ConvertToInt32(0);
-            sys_role.Enabled = Request.Form["select_enabled"].ConvertToInt32(0);
-            sys_role.CreateUserId = 0;
-            sys_role.CreateBy = "admin";
+            sys_role.Enabled = 1;
+            sys_role.ModifiedOn = DateTime.Now;
+            sys_role.ModifiedUserId = UserId.ConvertToInt32(0);
+            sys_role.ModifiedBy = UserName;
             int rtv = SysRoleDB.EditHandle(sys_role, string.Empty);
             string powers = Request.Form["hid_power"].ConventToString(string.Empty);
             if (!string.IsNullOrEmpty(powers))
@@ -173,13 +207,16 @@ namespace SunacCADApp.Controllers
                     Sys_Role_Model_Relation RoleModel = new Sys_Role_Model_Relation();
                     RoleModel.Model_Id = power.ConvertToInt32(-1);
                     RoleModel.Role_Id = Id;
-                    RoleModel.CreateOn = DateTime.Now;
                     RoleModel.Enabled = 1;
                     RoleModel.Reorder = 0;
+                    RoleModel.CreateUserId = UserId.ConvertToInt32(0);
+                    RoleModel.CreateBy = UserName;
+                    RoleModel.CreateOn = DateTime.Now;
+                    RoleModel.ModifiedUserId = UserId.ConvertToInt32(0);
+                    RoleModel.ModifiedBy = UserName;
                     RoleModel.ModifiedOn = DateTime.Now;
                     SysRoleModelRelationDB.AddHandle(RoleModel);
                 }
-
             }
 
             if (rtv > 0)
@@ -199,11 +236,14 @@ namespace SunacCADApp.Controllers
         /// <author>alon<84789887@qq.com></author>  
         public ActionResult DeleteHandleById()
         {
+            if (IUserId < 1)
+            {
+                return Redirect("/home");
+            }
             int Id = Request.QueryString["id"].ConvertToInt32(0);
             int rtv = SysRoleDB.DeleteHandleById(Id);
             string _wh = string.Format(" Role_Id='{0}'",Id);
             SysRoleModelRelationDB.DeleteHandleByParam(_wh);
-
             if (rtv > 0)
             {
                 return Json(new { code = 100, message = "删除成功" }, JsonRequestBehavior.AllowGet);
@@ -221,6 +261,10 @@ namespace SunacCADApp.Controllers
         /// <author>alon<84789887@qq.com></author> 
         public ActionResult DeleteHandleByIds()
         {
+            if (IUserId < 1)
+            {
+                return Redirect("/home");
+            }
             string ids = Request.QueryString["ids"].ConventToString(string.Empty);
             int rtv = SysRoleDB.DeleteHandleByIds(ids);
             if (rtv > 0)
